BitmapInfo::GetAnimationLayout(...) use an int instead of a size_t. Will be more consistent with other access with ivec2

+++ executables/LUDUMDARE/Ludum45/src/Ludum45Particles.h
@@ -136,7 +136,7 @@ public:
 
 	float  orientation = 0.0f;
 	float  image_timer = 0.0f;
-	size_t current_frame = 0;
+	int current_frame = 0;
 	
 
 };
